Implement FIDE algebraic notation interpretation
Define and implement the function Game::interpretCmd(), which
takes a string and interprets it as a move denoted in FIDE SAN.
Define a move type to hold this information, and throw an error
message if input is formatted incorrectly. Further define the
checkGameStatus() and updateBoard() functions to create a
program structure to be implemented in the future.