Add scripts to generate magic numbers and add a build directory
Create a python script "mine_magics.py" that implements an
algorithm to generate a series of "magic" numbers and a hash
map to be used for move generation. Edit both mine_magics.py
and generate_masks.py to write directly to "masks.cpp"
and "magics.cpp" respectively; doing this allows for future
editors to edit the generated magics and masks more easily.
Create a build directory to store object files in, and edit
the Makefile to reflect this to organize the repo files.