Add incremental updating to the board hashing function
Rewrite and refactor the board hashing function Board::GetBoardHash() so
that it simply acts as a getter member function, with the board hash
being incrementally updated as parts of the board position change; this
will increase the efficiency of Board::GetBoardHash() by avoiding
unnecessary XOR operations. Refactor ParseMoveCmd() for code readability
and fix a pawn promotion bug.