Cache the board state before iterative deepening
Change the implementation of iterative deepening to reset the board
state using data cached before search begins rather than forking the
process. This eliminates the need to copy data between processes while
also preventing problems with a partially manipulated board state that
occurs when an in-progress search is stopped when the time limit is
reached.